refactor: extract Adapter interface and registry out of cli.ts
Splits two concerns off of cli.ts:
- src/adapter/adapter.ts — the canonical Adapter contract, imported
  by cli and by future adapter authors as a named interface.
- src/adapter/registry.ts — the known-agent map and the Agent type.

cli.ts now imports both and keeps only the pipeline (run, dispatch,
safeEvaluate) plus the Agent type re-export for bin.ts.
